Moved title struct to its own header file
