+++ include/nspi_menu.h
@@ -6,19 +6,9 @@
 #include <vector>
 
 #include "nspi_pad.h"
+#include "nspi_title.h"
 
 namespace nspi {
-
-struct Title {
-  uint32_t releaseDate;    // 4 bytes
-  uint32_t size;           // 4 bytes
-  char region[3];          // 3 bytes
-  char id[17];             // 17 bytes
-  char publisher[25];      // 25 bytes
-  char name[49];           // 49 bytes
-  char version[49];        // 49 bytes
-  char description[1025];  // 1025 bytes
-};
 
 class Menu {
  private:
@@ -30,11 +20,10 @@ class Menu {
 
   uint16_t focusIndex;
   uint16_t focusOffset;
+  std::set<uint16_t> marked;
 
   Pad& pad;
   std::vector<Title> dummyData;
-
-  std::set<uint16_t> marked;
 
   void printHeader() const;
   void printContent() const;
